requirements.txt
python-docx
pypdf
PyQt5
whoosh
